test(postgres-source): extract PostgresSourceTestBase for shared CDC test plumbing
Pure extraction ahead of adding a second CDC test (a simulation test ported
from Clojure) that needs the same plumbing. Moves the container, pg writes,
node open/attach, cdc-db query, and the await helpers (including the
snapshotCompleted-token awaitStreaming) into a base class; the property test
keeps only its generators, normalisation, and assertions.

The container becomes a singleton — started once and reaped by Ryuk — rather
than per-class start/stop. A shared static container with inherited @BeforeAll
/@AfterAll would have the first subclass stop the instance the second still
needs; the singleton is the idiomatic testcontainers way to share across
classes and lets the whole module's CDC tests reuse one Postgres.
